python-ocr/requirements.txt
easyocr
fastapi
uvicorn
pillow
numpy
python-multipart
opencv-python-headless
pymupdf
python-docx
pdf2docx
pdfplumber
requests
pyspellchecker
docx2pdf
pywin32; platform_system == "Windows"
